Make JIRA an easy target for OOM killer
Eggplant occasionally dies with OOM situation. This just happened
this morning, and looking at syslog, I see that the OOM killer failed
to kill the real memory leak culprit, which is either JIRA or
Confluence.

This change will make it an easy target for killing, to prevent
system instability that prevents us from logging in.