don't let local modifications from blocking puppet apply.
If a local workspace is modified intentionally or unintentionally,
"git pull" will refuse to overwrite local changes with what's in
the repo, and puppet application will block forever.

It'd be much better to automatically correct this situation,
especially given that we have no ability to detect failures
to apply puppet configuration currently.

This has a risk of overwriting local changes anyone is making
on that box, but one is not supposed to do that anyway!